test(deepseek): move V4 thinking-mode tests to tests/test_litellm for CI coverage
Codecov patch coverage runs the tests/test_litellm suite; the new tests
lived in tests/litellm and were not executed there. Also adds an
async_transform_request test.
